requirements.txt
ccxt>=2.0.0
requests>=2.26.0
schedule>=1.1.0
python-dotenv>=0.19.0
pytz>=2021.1
python-telegram-bot==20.8 